backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.2.1
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.3
face_recognition==1.3.0
numpy==1.26.4
Pillow==10.2.0
openpyxl==3.1.2
aiofiles==23.2.1
